requirements/prod.txt
Flask
Flask-Bcrypt
Flask-Migrate
Flask-RESTful
Flask-SQLAlchemy
PyYAML
mysql-connector-python
